Add config.loadLoggersAndHandlers()
Add the function that actually calls the config parser and loads all the
loggers and handlers from configuration. Added a test to cover this
function, using both the environment variable and an input.

Because this required another JSON file for testing (although an
existing one could be reused or altered), renamed the directory where
these JSONs for tests are stored from configParserJsons to
testLoggingConfigurations.