Add JSON Schema for Logging, Loggers and Handlers
Add the JSON schemas for the main logging configuration, and two schemas
for the logger and hander specific configuration. These schemas are very
basic, but get the current job done. Some more advances schema settings
(like hiding properties that are for ConsoleHandler is the handler type
is FileHandler and vice versa) are for a later release.