Change configParser and Test for Schema Change
Changed all the test configurations to adhere to the JSON schema. Some
of the invalid test cases do violate that schema. For those,
"$comment"'s are added to make sure that nobody fixes those.

Update configParser to also adhere to the new JSON schema.

Some of the next steps might be to make the schema more solid, and use a
library that can use a JSON schema and validate against it. Saves me
maintaining a parser and a schema.