Use configuration parser package
Remove the configParser module that had a custom configuration parser.
That module inspired building a dedicated package to parse
configurations. This commit now uses that package. There are still some
enhancements required on the config parse package. For now there are
TODOs in the config module that explain what still needs to be validated
here, but adding more features to the config parser package will make
those validations obsolete.